+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * `Unreleased` - [View Diff](https://github.com/cortiz/prawn-rails/compare/v1.4.2...master)
   - [#52](https://github.com/cortiz/prawn-rails/pull/52) - Add active_support as a dependency and use the active_support lazy load hooks
+  - [#51](https://github.com/cortiz/prawn-rails/pull/51) - Remove usage of openstruct for config object
   - Remove unnecessary method `get_metadata` from `PrawnRailsHelper` which was getting included into `ActionView::Base`
   - Remove unnecessary rescue in `PrawnRails.config`
   - Remove redundant method `PrawnRails::Document.extensions`